Clear more sphagetti code and create a new module
